WIP on exporter class

+++ src/Prontotype/Utils.php
@@ -48,11 +48,11 @@ Class Utils {
         return file_exists($this->app['pt.prototype.paths.templates'] . '/' . $templatePath);
     }
     
-    protected function forcefileContents($path, $contents)
+    public function forcefileContents($path, $contents)
     {
         $file = basename($path);
         $dir = dirname($path);
-
+        
         if ( ! is_dir($dir) ) {
             mkdir($dir, 0771, true);
         }
@@ -61,16 +61,20 @@ Class Utils {
         chmod($path, 0644);
     }
 
-    protected function forceRemoveDir($dir)
+    public function forceRemoveDir($dir, $includeParent = true)
     {
-        foreach ( glob($dir . '/*') as $file ) {
-            if ( is_dir($file) ) {
-                $this->rrmdir( $file );
-            } else {
-                unlink($file);
+        if ( ! empty($dir) && $dir !== '/' ) {
+            foreach ( glob($dir . '/*') as $file ) {
+                if ( is_dir($file) ) {
+                    $this->forceRemoveDir( $file, true );
+                } else {
+                    unlink($file);
+                }
+            }
+            if ( $includeParent ) {
+                rmdir($dir);
             }
         }
-        rmdir($dir);
     }
     
 }
